pyproject.toml
[project]
name = "sql-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.19",
    "langchain-community>=0.3.18",
    "langchain-core>=0.3.37",
    "langchain-groq>=0.2.4",
    "langchain-ollama>=0.2.3",
    "langchain-openai>=0.3.6",
    "langgraph>=0.2.74",
    "openai>=1.63.2",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.0.1",
]
